split builds tdlibs
